fix: bring back list renderer
