+++ frappe/website/path_resolver.py
@@ -6,8 +6,8 @@ from werkzeug.routing import Rule
 
 import frappe
 from frappe.website.page_renderers.document_page import DocumentPage
+from frappe.website.page_renderers.list_renderer import ListPage
 from frappe.website.page_renderers.not_found_page import NotFoundPage
-from frappe.website.page_renderers.portal_renderer import PortalPage
 from frappe.website.page_renderers.print_page import PrintPage
 from frappe.website.page_renderers.redirect_page import RedirectPage
 from frappe.website.page_renderers.static_page import StaticPage
@@ -61,7 +61,7 @@ class PathResolver:
 			DocumentPage,
 			TemplatePage,
 			PrintPage,
-			PortalPage,
+			ListPage,
 		]
 
 		for renderer in renderers:
@@ -137,7 +137,6 @@ def resolve_redirect(path, query_string=None):
 		return
 
 	redirects = frappe.get_hooks("website_redirects")
-	redirects += add_portal_redirect()
 	redirects += [
 		{
 			"source": r.source,
@@ -229,20 +228,3 @@ def get_website_rules():
 def validate_path(path: str):
 	if not PathResolver(path).is_valid_path():
 		frappe.throw(frappe._("Path {0} it not a valid path").format(frappe.bold(path)))
-
-
-def add_portal_redirect():
-	redirects = []
-	menu_items = frappe.get_single("Portal Settings").menu
-
-	for item in menu_items:
-		if check_if_webform_exists(item.route):
-			continue
-		redirects.append(
-			{
-				"source": item.route,
-				"target": f"/portal{item.route}",
-				"forward_query_parameters": True,
-			}
-		)
-	return redirects
